+++ src/lib/commandBarConfigs/modelingCommandConfig.ts
@@ -1,5 +1,5 @@
 import { Models } from '@kittycad/lib'
-import { angleLengthInfo } from 'components/Toolbar/setAngleLength'
+import { angleLengthInfo } from 'components/Toolbar/angleLengthInfo'
 import { transformAstSketchLines } from 'lang/std/sketchcombos'
 import {
   isPathToNode,
