+++ src/machines/modelingMachine.ts
@@ -40,7 +40,7 @@ import {
   horzVertDistanceInfo,
 } from 'components/Toolbar/SetHorzVertDistance'
 import { angleBetweenInfo } from 'components/Toolbar/SetAngleBetween'
-import { angleLengthInfo } from 'components/Toolbar/setAngleLength'
+import { angleLengthInfo } from 'components/Toolbar/angleLengthInfo'
 import {
   applyConstraintEqualLength,
   setEqualLengthInfo,
